perf: Replace std::ostringstream with std::string concatenation
Replaced the use of `std::ostringstream` with a pre-allocated
`std::string` (`reserve()`) and `std::to_string()` concatenation
in `mutation_telemetry_json`. `std::ostringstream` involves substantial
formatting and dynamic allocation overhead which degrades performance
when serializing telemetry frequently in hot paths.

Also cast the `uint64_t` properties from `MutationTelemetrySnapshot`
to `unsigned long long` before passing them to `std::to_string()` to
ensure type portability issues are avoided on some platforms,
fixing the CI build error.

+++ src/telemetry/mutation_events.cpp
@@ -192,11 +192,11 @@ std::string mutation_telemetry_json(const MutationTelemetrySnapshot &snapshot) {
   std::string out;
   out.reserve(256);
   out += "{\"recipe_hits\":";
-  out += std::to_string(snapshot.recipe_hits);
+  out += std::to_string(static_cast<unsigned long long>(snapshot.recipe_hits));
   out += ",\"recipe_misses\":";
-  out += std::to_string(snapshot.recipe_misses);
+  out += std::to_string(static_cast<unsigned long long>(snapshot.recipe_misses));
   out += ",\"mutation_count\":";
-  out += std::to_string(snapshot.mutation_count);
+  out += std::to_string(static_cast<unsigned long long>(snapshot.mutation_count));
   out += ",\"operators\":{";
   bool first = true;
   for (const auto &[op, count] : snapshot.operator_counts) {
@@ -207,7 +207,7 @@ std::string mutation_telemetry_json(const MutationTelemetrySnapshot &snapshot) {
     out += "\"";
     out += json_escape(op);
     out += "\":";
-    out += std::to_string(count);
+    out += std::to_string(static_cast<unsigned long long>(count));
   }
   out += "}}";
   return out;
